refactor: Rename `sdk-js/src/lib/event-detection/eventTypes`
