+++ sdk-js/src/lib/event-detection/event-types/alexaSkill.js
