+++ sdk-js/src/lib/event-detection/event-types/apiGateway.js
