+++ sdk-js/src/lib/event-detection/event-types/apiGatewayV2.js
