+++ sdk-js/src/lib/event-detection/event-types/cloudwatchEvent.js
