+++ sdk-js/src/lib/event-detection/event-types/cloudwatchLog.js
