+++ sdk-js/src/lib/event-detection/event-types/customAuthorizer.js
