+++ sdk-js/src/lib/event-detection/event-types/dynamodb.js
