+++ sdk-js/src/lib/event-detection/event-types/firehose.js
