+++ sdk-js/src/lib/event-detection/event-types/kinesis.js
