+++ sdk-js/src/lib/event-detection/event-types/s3.js
