+++ sdk-js/src/lib/event-detection/event-types/ses.js
