+++ sdk-js/src/lib/event-detection/event-types/slsIntegrationLambda.js
