+++ sdk-js/src/lib/event-detection/event-types/sns.js
