+++ sdk-js/src/lib/event-detection/event-types/sqs.js
